// my-openapi-zod-app/src/api/Schema/ZDataMigrationThroughApiResponseModel.ts
//
// This file is generated. Do not edit manually.
//
// @applicationName: GreenOnion.API
// @applicationDescription: GreenOnion.API - 1.0.0.0
// @applicationVersion: 1.0.0.0
// @generatedDate: 2025/07/26
//

import { z } from "zod";

export const ZDataMigrationThroughApiResponseModel = z.object({
    success:z.boolean().optional(),
    message:z.string().optional(),
    totalRecords:z.number().int().optional(),
    payload:z.unknown().optional(),
    isUpdate:z.boolean().optional(),
    isAdded:z.boolean().optional(),
    isDelete:z.boolean().optional(),
});